requirements.txt
flask~=2.3.0
notion-client
pandas
python-dotenv~=1.1.1
pyyaml
requests~=2.32.5
ijson
psycopg2-binary
requests